The "show" command honors the git pager settings
This was a bit tricky (and not all that portable) because Git
apparently doesn't make its command-line available to the rits
script.  So instead, the script obtains the Git PID by assuming
git is the process group leader and then looking in /proc for
the command line. Let's hope it works well enough.